// src/app/types/booking.ts
export interface IBooking{
  participants: string;
  travelDate: string;
  roomCategory: string;
  shipName: string;
  firstName: string;
  surname: string;
  gender: string;
  dob: string;
  mobile: string;
  email: string;
  street: string;
  country: string;
  postcode: string;
  city: string;
  emergencyName: string;
  emergencyPhone: string;
  travelInsurance: string;
  insuranceProvider: string;
  policyNumber: string;
  additionalNote: string;
}